Try to use .js to avoid module import error
ERROR in ./node_modules/@tony/oembed-component/dist/o-embed.js 21:0-59
Module not found: Error: Can't resolve 'lit-html/directives/if-defined' in 'node_modules/@tony/oembed-component/dist'
Did you mean 'if-defined.js'?
BREAKING CHANGE: The request 'lit-html/directives/if-defined' failed to resolve only because it was resolved as fully specified
(probably because the origin is a '*.mjs' file or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.